remove if link exists

+++ scripts/stop.sh
@@ -1,1 +1,5 @@
-unlink /var/www/app
+#unlink /var/www/app
+if [[ -L /var/www/app ]]
+then
+    unlink /var/www/app 
+fi
